// src/app/api/models/translation-payload-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TranslationPayloadMeta } from '../models/translation-payload-meta';

/**
 * Response DTO containing translation payload for a specific language
 */
export interface TranslationPayloadDto {

  /**
   * Nested JSON translation data
   */
  data?: {
[key: string]: any;
};

  /**
   * Metadata about the translation payload
   */
  meta?: TranslationPayloadMeta;
}