// src/app/api/models/language-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Language metadata DTO
 */
export interface LanguageDto {

  /**
   * Language code
   */
  code?: string;

  /**
   * Text direction: "ltr" or "rtl"
   */
  dir?: string;

  /**
   * Whether this language is active
   */
  isActive?: boolean;

  /**
   * English name of the language
   */
  name?: string;

  /**
   * Native name of the language
   */
  nativeName?: string;
}